==> 00-importacao_e_estruturacao.ipynb <==
# Descrição dos dados (do Kaggle):

In this competition you will develop algorithms to classify genetic mutations based on clinical evidence (text).

There are nine different classes a genetic mutation can be classified on.

This is not a trivial task since interpreting clinical evidence is very challenging even for human specialists. Therefore, modeling the clinical evidence (text) will be critical for the success of your approach.

Both, training and test, data sets are provided via two different files. One (training/test_variants) provides the information about the genetic mutations, whereas the other (training/test_text) provides the clinical evidence (text) that our human experts used to classify the genetic mutations. Both are linked via the ID field.

Therefore the genetic mutation (row) with ID=15 in the file training_variants, was classified using the clinical evidence (text) from the row with ID=15 in the file training_text

Finally, to make it more exciting!! Some of the test data is machine-generated to prevent hand labeling. You will submit all the results of your classification algorithm, and we will ignore the machine-generated samples. 

File descriptions
- training_variants - a comma separated file containing the description of the genetic mutations used for training. Fields are ID (the id of the row used to link the mutation to the clinical evidence), Gene (the gene where this genetic mutation is located), Variation (the aminoacid change for this mutations), Class (1-9 the class this genetic mutation has been classified on)
- training_text - a double pipe (||) delimited file that contains the clinical evidence (text) used to classify genetic mutations. Fields are ID (the id of the row used to link the clinical evidence to the genetic mutation), Text (the clinical evidence used to classify the genetic mutation)
- test_variants - a comma separated file containing the description of the genetic mutations used for training. Fields are ID (the id of the row used to link the mutation to the clinical evidence), Gene (the gene where this genetic mutation is located), Variation (the aminoacid change for this mutations)
- test_text - a double pipe (||) delimited file that contains the clinical evidence (text) used to classify genetic mutations. Fields are ID (the id of the row used to link the clinical evidence to the genetic mutation), Text (the clinical evidence used to classify the genetic mutation)
- submissionSample - a sample submission file in the correct format

# Neste caso, o dataset de test não fornece a classe de mutação, pois é destinado somente à previsão para
# submissão à competição do kaggle.

# In this case, the test dataset does not provide the class of the mutation. The test dataset is only for
# submition to kaggle.

## Carregando o arquivo training_text

import pandas as pd

## Função para estruturar o texto bruto em um dicionário, salvo como resultado intermediário em um artuivo json.

## Function to give the raw text a structure and save the intermediary result in a json file.

def structure_text(path, json_name):
    f = open(path, 'r')
    text = f.read()
    f.close()
    
    lines = text.split(sep='\n')
    dummy = {'ID':[], 'Text':[]}
    
    for line in lines:
        splitted_line = line.split(sep='||')
        if len(splitted_line) > 1:
            dummy['ID'].append(splitted_line[0])
            dummy['Text'].append(splitted_line[1])
    
    df = pd.DataFrame.from_dict(dummy)
    df.set_index('ID', inplace=True)
    df.to_json(path_or_buf=json_name, orient='split')

path_1 = './data_files/training_text'
#path_2 = './data_files/test_text'

structure_text(path_1, './data_files/training_text.json')

#structure_text(path_2, './data_files/test_text.json')

df1 = pd.read_json('./data_files/training_text.json', orient='split')
df1.head()

#df2 = pd.read_json('./data_files/test_text.json', orient='split')
#df2.head()

df1

## Carregando o arquivo training_variants

df2 = pd.read_csv('./data_files/training_variants')

df2

df2.set_index('ID', inplace=True)
df2

# Salvando resultado intermediário in json
# Saving intermediary result in json file.

df2.to_json(path_or_buf='./data_files/training_variants.json', orient='split')

df1

## Fazendo a união dos dois datasets pelo índice

df3 = df1.merge(df2, left_index=True, right_index=True)
df3
