==> 01-Analise_exploratoria_de_dados.ipynb <==
# Análise Exploratória de Dados

A partir do dataset formatado na etapa anterior, será feita uma análise exploratória dos dados

import pandas as pd

df = pd.read_json('./data_files/merged_datasets.json', orient='split')
df

# Verificando valores NA's
df.isna().sum()

# Verificando os tipos das variáveis
df.dtypes

# Verificando valores únicos
df.nunique()

Notar que enquanto o dataset possui 3221 observações, há apenas 1921 textos distintos e 2996 valores de Variation distintos. Isto significa que pode haver mais de uma variação para cada texto descritivo.

# Verificando se há linhas duplicadas
df.duplicated().sum()

Através de inspeção no arquivo, o atributo Variation traz o nome da variação, com exceção de algumas classes comuns como deleção, fusão, etc. Deste modo, conclui-se que não acrescenta informação relevante à análise e será excluída do dataset.

df.drop('Variation', inplace=True, axis=1)

df

# Salvando o resultado intermediário em um json
df.to_json('./data_files/df_text_gene_class.json')

df.groupby(by='Class').count()

Pode-se observar um desbalanceamento do dataset, com as classes 3, 8 e 9 com poucas observações. Posteriormente, pode ser usada alguma técnica de balanceamento como SMOTE para gerar dados sintéticos
