==> 03-Limpeza_de_dados.ipynb <==
# Engenharia de atributos
Nesta etapa será realizada a limpeza dos dados gerados pelo cálculo do TF-IDF, bem como balanceamento de classes e redução de dimensionalidade do dataset. Ao final, é gerada a matriz de treinamento do modelo de machine learning.

import pandas as pd
import numpy as np

# Carga do arquivo completo de TF-IDF
# Não efetuada por causa do problema de estouro de memória da máquina local
#df1 = pd.read_json('./data_files/data_tf_idf_completo.json', orient='column')

df2 = pd.read_json('./data_files/data_tf_idf.json', orient='column')

df2.shape

df2.head()

df = df2.T
del df2

df.shape

df.head()

## Tratamento de NA's e normalização

df['predicted'].notna()

# Preenchendo os NA's com o valor zero, dado que correspondem a palavras que não foram encontradas nos docs.
df.fillna(0, inplace=True)

df.head()

tf_idf = df.values

type(tf_idf)

tf_idf.shape

tf_idf.max()

# Normalização dos dados pelo maior valor de TF_IDF
tf_idf = tf_idf / tf_idf.max()

tf_idf

tf_idf.max()

del df

## Divisão em dados de treino e teste

import sklearn
from sklearn.model_selection import train_test_split

df = pd.read_json('data_files/merged_datasets.json', orient='split')
df.head()

# Contagem de observações por classe
df.groupby(by='Class').count()

 Indica necessidade de balanceamento antes de treinamento do modelo. Porém, antes deve-se fazer a divisão em dados de treino e teste.

y = df['Class'].values
y

y.shape

X_train, X_test, y_train, y_test = train_test_split(tf_idf, y, test_size=0.3)

# Verificando as dimensões:

dict = {'X_train': X_train.shape, 'X_test': X_test.shape, 'y_train': y_train.shape, 'y_test': y_test.shape}

dict

del df

# Salvando os arrays para balanceamento em outro notebook

np.savetxt('data_files/X_train.csv', X_train, delimiter=',')
np.savetxt('data_files/y_train.csv', y_train, delimiter=',')
np.savetxt('data_files/X_test.csv', X_test, delimiter=',')
np.savetxt('data_files/y_test.csv', y_test, delimiter=',')
