==> 05-Reducao_dimensionalidade.ipynb <==
# Redução de dimensionalidade
Nesta etapa será feita a redução de dimensionalidade antes do treinamento dos modelos de machine learning.

import numpy as np
from sklearn.decomposition import SparsePCA
from scipy import sparse

from google.colab import drive

drive.mount('/content/drive')

!ls -la drive/MyDrive/'Colab Notebooks'/

X_train_blc_sparse = sparse.load_npz('drive/MyDrive/''Colab Notebooks''/X_train_blc_sparse.npz')

type(X_train_blc_sparse)

X_train_balanced = X_train_blc_sparse.toarray()

type(X_train_balanced)

X_train_balanced.shape

## Decomposição com o método PCA
Foi aplicado o método de *principal component analysis (PCA)* para redução de dimensionalidade do dataset de treino.

pca = SparsePCA(n_components=500)

X_train_pca = pca.fit_transform(X_train_balanced)

X_train_pca.shape

# Mesma sequência de operações com o dataset de teste

X_test_sparse = sparse.load_npz('drive/MyDrive/''Colab Notebooks''/X_test_sparse.npz')

X_test = X_test_sparse.toarray()

type(X_test)

X_test.shape

X_test_pca = pca.fit_transform(X_test)

X_test_pca.shape

## Salvando os resultados de X_train_pca e X_test_pca para entrada no treinamento dos modelos de ML

np.savetxt('drive/MyDrive/''Colab Notebooks''/X_train_pca.csv', X_train_pca, delimiter=',')

np.savetxt('drive/MyDrive/''Colab Notebooks''/X_test_pca.csv', X_test_pca, delimiter=',')
